+++ multiboot2/src/boot_information.rs
@@ -8,16 +8,19 @@ use crate::{
     ElfSectionIter, ElfSectionsTag, EndTag, FramebufferTag, ImageLoadPhysAddrTag, MemoryMapTag,
     ModuleIter, RsdpV1Tag, RsdpV2Tag, SmbiosTag, TagIter, TagType, VBEInfoTag,
 };
+#[cfg(feature = "unstable")]
+use core::error::Error;
 use core::fmt;
 use core::mem;
 use core::ptr::NonNull;
 use derive_more::Display;
 use multiboot2_common::{DynSizedStructure, Header, MaybeDynSized, MemoryError, Tag};
 
-/// Error type that describes errors while loading/parsing a multiboot2 information structure
-/// from a given address.
+/// Errors that occur when a chunk of memory can't be parsed as
+/// [`BootInformation`].
 #[derive(Display, Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum MbiLoadError {
+pub enum LoadError {
+    /// The provided memory can't be parsed as [`BootInformation`].
     /// See [`MemoryError`].
     Memory(MemoryError),
     /// Missing mandatory end tag.
@@ -25,7 +28,14 @@ pub enum MbiLoadError {
 }
 
 #[cfg(feature = "unstable")]
-impl core::error::Error for MbiLoadError {}
+impl Error for LoadError {
+    fn source(&self) -> Option<&(dyn Error + 'static)> {
+        match self {
+            Self::Memory(inner) => Some(inner),
+            Self::NoEndTag => None,
+        }
+    }
+}
 
 /// The basic header of a [`BootInformation`] as sized Rust type.
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -91,13 +101,13 @@ impl<'a> BootInformation<'a> {
     ///   boot environments, such as UEFI.
     /// * The memory at `ptr` must not be modified after calling `load` or the
     ///   program may observe unsynchronized mutation.
-    pub unsafe fn load(ptr: *const BootInformationHeader) -> Result<Self, MbiLoadError> {
-        let ptr = NonNull::new(ptr.cast_mut()).ok_or(MbiLoadError::Memory(MemoryError::Null))?;
-        let inner = DynSizedStructure::ref_from_ptr(ptr).map_err(MbiLoadError::Memory)?;
+    pub unsafe fn load(ptr: *const BootInformationHeader) -> Result<Self, LoadError> {
+        let ptr = NonNull::new(ptr.cast_mut()).ok_or(LoadError::Memory(MemoryError::Null))?;
+        let inner = DynSizedStructure::ref_from_ptr(ptr).map_err(LoadError::Memory)?;
 
         let this = Self(inner);
         if !this.has_valid_end_tag() {
-            return Err(MbiLoadError::NoEndTag);
+            return Err(LoadError::NoEndTag);
         }
         Ok(this)
     }